Fixed erros, added variablres for clarity
Fixed error regardin bno class, where variables where missused, added variables in strategy to clarify the use of the pixy coordinates. Fixed minor errors and typos.

+++ motor.h
@@ -1,3 +1,5 @@
+#include <Arduino.h>
+
 class Motor
 {
 public:
@@ -6,11 +8,11 @@ public:
     void forward();             //Sets motor to spin forward.
     void backward();            //Sets motor to spin backards.
     void stop();                //Makes motor stop spinning.
-    void changePwm(float data); //Changes the pwm of the motor.
+    void changePwm(int data); //Changes the pwm of the motor.
 
 private:
     int motorBackwards;
     int motorForward;
     int analogPin;
-    float pwm = 0;
+    int pwm = 0;
 };
